Moved searchStations function into funcs folder
